Consolidate the various stack traits.
There is only one type (Stack) intended to implement the traits
StackEvaluate, StackKeys, StackMatches, and StackProgram. These traits
exist so that `synthesizer_program` can deal with stacks,
but the concrete implementation can be in `synthesizer_process`,
since it depends on items there. These four traits have now been
consolidated into one, StackTrait.

The traits StackEvaluate, StackExecute, and StackProgramTypes
served no purpose as separate traits and their functions are now
in Stack.

Some functions in `synthesizer_process` now accept a concrete `Stack`.
