+++ synthesizer/process/src/stack/call/mod.rs
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{CallStack, Registers, RegistersCall, StackEvaluate, StackExecute, stack::Address};
+use crate::{CallStack, Registers, RegistersCall, Stack, stack::Address};
 use aleo_std::prelude::{finish, lap, timer};
 use console::{
     account::Field,
@@ -30,23 +30,17 @@ use synthesizer_program::{
     RegistersSignerCircuit,
     RegistersStore,
     RegistersStoreCircuit,
-    StackKeys,
-    StackMatches,
-    StackProgram,
+    StackTrait,
 };
 
 pub trait CallTrait<N: Network> {
     /// Evaluates the instruction.
-    fn evaluate<A: circuit::Aleo<Network = N>>(
-        &self,
-        stack: &(impl StackEvaluate<N> + StackMatches<N> + StackProgram<N>),
-        registers: &mut Registers<N, A>,
-    ) -> Result<()>;
+    fn evaluate<A: circuit::Aleo<Network = N>>(&self, stack: &Stack<N>, registers: &mut Registers<N, A>) -> Result<()>;
 
     /// Executes the instruction.
     fn execute<A: circuit::Aleo<Network = N>, R: CryptoRng + Rng>(
         &self,
-        stack: &(impl StackEvaluate<N> + StackExecute<N> + StackMatches<N> + StackKeys<N> + StackProgram<N>),
+        stack: &Stack<N>,
         registers: &mut (
                  impl RegistersCall<N>
                  + RegistersSigner<N>
@@ -61,11 +55,7 @@ pub trait CallTrait<N: Network> {
 impl<N: Network> CallTrait<N> for Call<N> {
     /// Evaluates the instruction.
     #[inline]
-    fn evaluate<A: circuit::Aleo<Network = N>>(
-        &self,
-        stack: &(impl StackEvaluate<N> + StackMatches<N> + StackProgram<N>),
-        registers: &mut Registers<N, A>,
-    ) -> Result<()> {
+    fn evaluate<A: circuit::Aleo<Network = N>>(&self, stack: &Stack<N>, registers: &mut Registers<N, A>) -> Result<()> {
         let timer = timer!("Call::evaluate");
 
         // Load the operands values.
@@ -143,7 +133,7 @@ impl<N: Network> CallTrait<N> for Call<N> {
     #[inline]
     fn execute<A: circuit::Aleo<Network = N>, R: Rng + CryptoRng>(
         &self,
-        stack: &(impl StackEvaluate<N> + StackExecute<N> + StackMatches<N> + StackKeys<N> + StackProgram<N>),
+        stack: &Stack<N>,
         registers: &mut (
                  impl RegistersCall<N>
                  + RegistersSigner<N>
