+++ synthesizer/process/src/stack/finalize_registers/load.rs
@@ -23,7 +23,7 @@ impl<N: Network> RegistersLoad<N> for FinalizeRegisters<N> {
     /// This method will halt if the register locator is not found.
     /// In the case of register accesses, this method will halt if the access is not found.
     #[inline]
-    fn load(&self, stack: &(impl StackMatches<N> + StackProgram<N>), operand: &Operand<N>) -> Result<Value<N>> {
+    fn load(&self, stack: &impl StackTrait<N>, operand: &Operand<N>) -> Result<Value<N>> {
         // Retrieve the register.
         let register = match operand {
             // If the operand is a literal, return the literal.
