+++ synthesizer/process/src/stack/finalize_registers/mod.rs
@@ -28,8 +28,7 @@ use synthesizer_program::{
     Operand,
     RegistersLoad,
     RegistersStore,
-    StackMatches,
-    StackProgram,
+    StackTrait,
 };
 
 use indexmap::IndexMap;
