+++ synthesizer/process/src/stack/register_types/mod.rs
@@ -16,6 +16,8 @@
 mod initialize;
 mod matches;
 
+use crate::Stack;
+
 use console::{
     network::prelude::*,
     program::{
@@ -41,8 +43,7 @@ use synthesizer_program::{
     Opcode,
     Operand,
     Program,
-    StackMatches,
-    StackProgram,
+    StackTrait,
 };
 
 use console::program::{FinalizeType, Locator};
@@ -60,14 +61,14 @@ impl<N: Network> RegisterTypes<N> {
     /// Initializes a new instance of `RegisterTypes` for the given closure.
     /// Checks that the given closure is well-formed for the given stack.
     #[inline]
-    pub fn from_closure(stack: &(impl StackMatches<N> + StackProgram<N>), closure: &Closure<N>) -> Result<Self> {
+    pub fn from_closure(stack: &Stack<N>, closure: &Closure<N>) -> Result<Self> {
         Self::initialize_closure_types(stack, closure)
     }
 
     /// Initializes a new instance of `RegisterTypes` for the given function.
     /// Checks that the given function is well-formed for the given stack.
     #[inline]
-    pub fn from_function(stack: &(impl StackMatches<N> + StackProgram<N>), function: &Function<N>) -> Result<Self> {
+    pub fn from_function(stack: &Stack<N>, function: &Function<N>) -> Result<Self> {
         Self::initialize_function_types(stack, function)
     }
 
@@ -86,11 +87,7 @@ impl<N: Network> RegisterTypes<N> {
     }
 
     /// Returns the register type of the given operand.
-    pub fn get_type_from_operand(
-        &self,
-        stack: &(impl StackMatches<N> + StackProgram<N>),
-        operand: &Operand<N>,
-    ) -> Result<RegisterType<N>> {
+    pub fn get_type_from_operand(&self, stack: &impl StackTrait<N>, operand: &Operand<N>) -> Result<RegisterType<N>> {
         Ok(match operand {
             Operand::Literal(literal) => RegisterType::Plaintext(PlaintextType::from(literal.to_type())),
             Operand::Register(register) => self.get_type(stack, register)?,
@@ -103,11 +100,7 @@ impl<N: Network> RegisterTypes<N> {
     }
 
     /// Returns the register type of the given register.
-    pub fn get_type(
-        &self,
-        stack: &(impl StackMatches<N> + StackProgram<N>),
-        register: &Register<N>,
-    ) -> Result<RegisterType<N>> {
+    pub fn get_type(&self, stack: &impl StackTrait<N>, register: &Register<N>) -> Result<RegisterType<N>> {
         // Initialize a tracker for the register type.
         let register_type = if self.is_input(register) {
             // Retrieve the input value type as a register type.
