+++ synthesizer/process/src/stack/registers/load.rs
@@ -22,7 +22,7 @@ impl<N: Network, A: circuit::Aleo<Network = N>> RegistersLoad<N> for Registers<N
     /// This method will halt if the register locator is not found.
     /// In the case of register accesses, this method will halt if the access is not found.
     #[inline]
-    fn load(&self, stack: &(impl StackMatches<N> + StackProgram<N>), operand: &Operand<N>) -> Result<Value<N>> {
+    fn load(&self, stack: &impl StackTrait<N>, operand: &Operand<N>) -> Result<Value<N>> {
         // Retrieve the register.
         let register = match operand {
             // If the operand is a literal, return the literal.
@@ -87,11 +87,7 @@ impl<N: Network, A: circuit::Aleo<Network = N>> RegistersLoadCircuit<N, A> for R
     /// This method will halt if the register locator is not found.
     /// In the case of register accesses, this method will halt if the access is not found.
     #[inline]
-    fn load_circuit(
-        &self,
-        stack: &(impl StackMatches<N> + StackProgram<N>),
-        operand: &Operand<N>,
-    ) -> Result<circuit::Value<A>> {
+    fn load_circuit(&self, stack: &impl StackTrait<N>, operand: &Operand<N>) -> Result<circuit::Value<A>> {
         use circuit::Inject;
 
         // Retrieve the register.
