+++ synthesizer/program/src/logic/command/remove.rs
@@ -17,7 +17,7 @@ use crate::{
     FinalizeOperation,
     Opcode,
     Operand,
-    traits::{FinalizeStoreTrait, RegistersLoad, StackMatches, StackProgram},
+    traits::{FinalizeStoreTrait, RegistersLoad, StackTrait},
 };
 use console::{network::prelude::*, program::Identifier};
 
@@ -62,7 +62,7 @@ impl<N: Network> Remove<N> {
     #[inline]
     pub fn finalize(
         &self,
-        stack: &(impl StackMatches<N> + StackProgram<N>),
+        stack: &impl StackTrait<N>,
         store: &impl FinalizeStoreTrait<N>,
         registers: &mut impl RegistersLoad<N>,
     ) -> Result<Option<FinalizeOperation<N>>> {
