+++ synthesizer/program/src/logic/instruction/operation/call.rs
@@ -16,7 +16,7 @@
 use crate::{
     Opcode,
     Operand,
-    traits::{RegistersLoad, RegistersLoadCircuit, StackMatches, StackProgram},
+    traits::{RegistersLoad, RegistersLoadCircuit, StackTrait},
 };
 use console::{
     network::prelude::*,
@@ -150,7 +150,7 @@ impl<N: Network> Call<N> {
 impl<N: Network> Call<N> {
     /// Returns `true` if the instruction is a function call.
     #[inline]
-    pub fn is_function_call(&self, stack: &impl StackProgram<N>) -> Result<bool> {
+    pub fn is_function_call(&self, stack: &impl StackTrait<N>) -> Result<bool> {
         match self.operator() {
             // Check if the locator is for a function.
             CallOperator::Locator(locator) => {
@@ -167,14 +167,14 @@ impl<N: Network> Call<N> {
     }
 
     /// Evaluates the instruction.
-    pub fn evaluate(&self, _stack: &impl StackProgram<N>, _registers: &mut impl RegistersLoad<N>) -> Result<()> {
+    pub fn evaluate(&self, _stack: &impl StackTrait<N>, _registers: &mut impl RegistersLoad<N>) -> Result<()> {
         bail!("Forbidden operation: Evaluate cannot invoke a 'call' directly. Use 'call' in 'Stack' instead.")
     }
 
     /// Executes the instruction.
     pub fn execute<A: circuit::Aleo<Network = N>>(
         &self,
-        _stack: &impl StackProgram<N>,
+        _stack: &impl StackTrait<N>,
         _registers: &mut impl RegistersLoadCircuit<N, A>,
     ) -> Result<()> {
         bail!("Forbidden operation: Execute cannot invoke a 'call' directly. Use 'call' in 'Stack' instead.")
@@ -182,11 +182,7 @@ impl<N: Network> Call<N> {
 
     /// Finalizes the instruction.
     #[inline]
-    pub fn finalize(
-        &self,
-        _stack: &(impl StackMatches<N> + StackProgram<N>),
-        _registers: &mut impl RegistersLoad<N>,
-    ) -> Result<()> {
+    pub fn finalize(&self, _stack: &impl StackTrait<N>, _registers: &mut impl RegistersLoad<N>) -> Result<()> {
         bail!("Forbidden operation: Finalize cannot invoke a 'call' directly. Use 'call' in 'Stack' instead.")
     }
 
@@ -194,7 +190,7 @@ impl<N: Network> Call<N> {
     #[inline]
     pub fn output_types(
         &self,
-        stack: &impl StackProgram<N>,
+        stack: &impl StackTrait<N>,
         input_types: &[RegisterType<N>],
     ) -> Result<Vec<RegisterType<N>>> {
         // Retrieve the external stack, if needed, and the resource.
